+++ database/migrations/2026_03_07_152232_create_opname_stock_items_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('opname_stock_id')->constrained()->cascadeOnDelete();
             $table->foreignId('item_id')->constrained()->cascadeOnDelete();
-            $table->integer('system_stock');
-            $table->integer('physical_stock');
-            $table->integer('difference');
+            $table->decimal('system_stock', 15, 2);
+            $table->decimal('physical_stock', 15, 2);
+            $table->decimal('difference', 15, 2);
             $table->timestamps();
         });
     }
