+++ database/migrations/2026_03_11_064230_create_invoice_items_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->foreignId('invoice_id')->constrained()->cascadeOnDelete();
             $table->foreignId('item_id')->constrained()->cascadeOnDelete();
 
-            $table->integer('quantity');
+            $table->decimal('quantity', 15, 2);
             $table->decimal('price', 15, 2);
             $table->decimal('total', 15, 2);
